refaactor handleclick function
