+++ src/components/editor/leftNav/ActivePanelPropertyNav.jsx
@@ -1,17 +1,15 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { setCurrentComponent } from "actions/index"
-import { useDispatch, useSelector } from "react-redux"
+import { useSelector } from "react-redux"
 import { displayResourceValidations } from "selectors/errors"
 import { selectNormProperty, selectNormValues } from "selectors/resources"
 import { selectPropertyTemplate } from "selectors/templates"
 import SubjectSubNav from "./SubjectSubNav"
 import PresenceIndicator from "./PresenceIndicator"
-import { selectModalType } from "selectors/modals"
+import useLeftNav from "hooks/useLeftNav"
 import _ from "lodash"
 
 const ActivePanelPropertyNav = (props) => {
-  const dispatch = useDispatch()
   const property = useSelector((state) =>
     selectNormProperty(state, props.propertyKey)
   )
@@ -21,6 +19,8 @@ const ActivePanelPropertyNav = (props) => {
   const propertyTemplate = useSelector((state) =>
     selectPropertyTemplate(state, property?.propertyTemplateKey)
   )
+
+  const handleClick = useLeftNav(property)
 
   const liClassNames = []
 
@@ -45,20 +45,6 @@ const ActivePanelPropertyNav = (props) => {
     return <ul>{subNavItems}</ul>
   }
 
-  const isModalOpen = useSelector((state) => selectModalType(state))
-
-  const handleClick = (property) => {
-    if (isModalOpen) return // do not respond to clicks in the nav if any modal is open
-
-    dispatch(
-      setCurrentComponent(
-        property.rootSubjectKey,
-        property.rootPropertyKey,
-        property.key
-      )
-    )
-  }
-
   if (!property) return null
 
   // Render this property and any children value subjects (if a property type = resource).
@@ -69,7 +55,7 @@ const ActivePanelPropertyNav = (props) => {
         className="btn btn-primary"
         aria-label={`Go to ${propertyTemplate.label}`}
         data-testid={`Go to ${propertyTemplate.label}`}
-        onClick={() => handleClick(property)}
+        onClick={handleClick}
       >
         <h5 className={headingClassNames.join(" ")}>
           {propertyTemplate.label}
