+++ src/components/editor/leftNav/PanelPropertyNav.jsx
@@ -1,16 +1,14 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { setCurrentComponent } from "actions/index"
-import { useDispatch, useSelector } from "react-redux"
+import { useSelector } from "react-redux"
 import { displayResourceValidations } from "selectors/errors"
 import { selectNormProperty } from "selectors/resources"
 import { selectPropertyTemplate } from "selectors/templates"
 import PresenceIndicator from "./PresenceIndicator"
-import { selectModalType } from "selectors/modals"
+import useLeftNav from "hooks/useLeftNav"
 import _ from "lodash"
 
 const PanelPropertyNav = (props) => {
-  const dispatch = useDispatch()
   const property = useSelector((state) =>
     selectNormProperty(state, props.propertyKey)
   )
@@ -25,21 +23,9 @@ const PanelPropertyNav = (props) => {
   const headingClassNames = ["left-nav-header"]
   if (displayValidations && hasError) headingClassNames.push("text-danger")
 
-  const isModalOpen = useSelector((state) => selectModalType(state))
+  const handleClick = useLeftNav(property)
 
   if (!property) return null
-
-  const handleClick = (property) => {
-    if (isModalOpen) return // do not respond to clicks in the nav if any modal is open
-
-    dispatch(
-      setCurrentComponent(
-        property.rootSubjectKey,
-        property.rootPropertyKey,
-        property.key
-      )
-    )
-  }
 
   return (
     <li>
@@ -48,7 +34,7 @@ const PanelPropertyNav = (props) => {
         className="btn btn-link"
         aria-label={`Go to ${propertyTemplate.label}`}
         data-testid={`Go to ${propertyTemplate.label}`}
-        onClick={() => handleClick(property)}
+        onClick={handleClick}
       >
         <h5 className={headingClassNames.join(" ")}>
           {propertyTemplate.label}
