+++ src/components/editor/leftNav/PropertySubNav.jsx
@@ -1,18 +1,15 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { setCurrentComponent } from "actions/index"
-import { useDispatch, useSelector } from "react-redux"
+import { useSelector } from "react-redux"
 import { displayResourceValidations } from "selectors/errors"
 import { selectNormProperty, selectNormValues } from "selectors/resources"
 import { selectPropertyTemplate } from "selectors/templates"
 import SubjectSubNav from "./SubjectSubNav"
 import PresenceIndicator from "./PresenceIndicator"
-import { selectModalType } from "selectors/modals"
+import useLeftNav from "hooks/useLeftNav"
 import _ from "lodash"
 
 const PropertySubNav = (props) => {
-  const dispatch = useDispatch()
-
   const property = useSelector((state) =>
     selectNormProperty(state, props.propertyKey)
   )
@@ -23,19 +20,7 @@ const PropertySubNav = (props) => {
     selectNormValues(state, property?.valueKeys)
   )
 
-  const isModalOpen = useSelector((state) => selectModalType(state))
-
-  const handleClick = (property) => {
-    if (isModalOpen) return // do not respond to clicks in the nav if any modal is open
-
-    dispatch(
-      setCurrentComponent(
-        property.rootSubjectKey,
-        property.rootPropertyKey,
-        property.key
-      )
-    )
-  }
+  const handleClick = useLeftNav(property)
 
   const hasError = !_.isEmpty(property.descWithErrorPropertyKeys)
   const displayValidations = useSelector((state) =>
@@ -65,7 +50,7 @@ const PropertySubNav = (props) => {
         className="btn btn-link"
         aria-label={`Go to ${propertyTemplate.label}`}
         data-testid={`Go to ${propertyTemplate.label}`}
-        onClick={() => handleClick(property)}
+        onClick={handleClick}
       >
         <span className={headingClassNames.join(" ")}>
           {propertyTemplate.label}
